backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
httpx>=0.25.0
python-dateutil>=2.8.0
selenium>=4.15.0
webdriver-manager>=4.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
